New API prototype signature corrected.

+++ CodeXL/Components/PowerProfiling/Backend/AMDTPowerProfilingDrivers/inc/AMDTHelpers.h
@@ -129,7 +129,7 @@ bool HelpAccessMSRAddress(PACCESS_MSR pData);
 uint64 HelpReadMsr64(uint32 reg);
 
 // HelpWriteMsr64: Write 64 bit MSR address
-uint64 HelpWriteMsr64(uint32 reg, uint64 value);
+uint32 HelpWriteMsr64(uint32 reg, uint64 value);
 
 // EncodeExtendedConfigSpaceAddress: Encode PCI address to extended PCI config speace.
 uint32 HelpEncodeExtendedConfigSpaceAddress(uint32 bus,
